eliminate need for simple registration plugin
Fixes #99

No longer need the plugin because we're back to showing sign-in and register at the same time (the default behavior).

Removed code where I was inserted the shortcode.

Copied over code from the plugin that was adding the First Name and Last Name fields to registration, and put that into the Luxcentric theme/plugin.